implemented betting, splitting

+++ global.h
@@ -1,6 +1,6 @@
 #ifndef GLOBAL_H
 #define GLOBAL_H
 
-extern int money;
+extern double money;
 
 #endif // GLOBAL_H
